NMR-6976: use TitlePaneSkin instead of binding width property, extend accordion in private inner class
